Маршрут загрузок
Теперь по маршруту uploader/my можно получить все загрузки текущего пользователя. Работают проекции.
